+++ solvers/heat_conduction_discrete_analogue.py
